+++ lsd_slam_core/src/Tracking/SE3Tracker.cpp
@@ -269,8 +269,8 @@ SE3 SE3Tracker::trackFrame(
 	if (plotTrackingIterationInfo)
 	{
 		const float* frameImage = frame->image();
-		for (int row = 0; row < model->h; ++ row)
-			for (int col = 0; col < model->w; ++ col)
+		for (size_t row = 0; row < model->h; ++ row)
+			for (size_t col = 0; col < model->w; ++ col)
 				setPixelInCvMat(&debugImageSecondFrame,getGrayCvPixel(frameImage[col+row*model->w]), col, row, 1);
 	}
